requirements.txt
Cython==0.20.1
Divisi2==2.2.5
Django==1.6.3
Fabric==1.10.0
Jinja2==2.7.3
MarkupSafe==0.23
South==1.0.1
Unidecode==0.04.16
csc-pysparse==1.1.1.4
csc-utils==0.6.7
decorator==3.4.0
dj-database-url==0.3.0
dj-static==0.0.5
django-bower==5.0.1
django-nvd3==0.7.7
django-toolbelt==0.0.1
ecdsa==0.11
gunicorn==18.0
networkx==1.8.1
numpy==1.9.1
paramiko==1.15.1
psycopg2==2.5.2
pycrypto==2.6.1
pystache==0.5.3
python-nvd3==0.13.5
python-recsys==0.2
python-slugify==0.0.9
scipy==0.13.3
six==1.8.0
sparsesvd==0.2.2
static==1.0.2
wsgiref==0.1.2
